Update 2016_09_13_070520_add_verification_to_user_table.php
Fixes a weird issue with SQLite, when rolling back the migration

"SQLSTATE[HY000]: General error: 1 no such column: verified"

+++ src/resources/migrations/2016_09_13_070520_add_verification_to_user_table.php
@@ -43,8 +43,20 @@ class AddVerificationToUserTable extends Migration
     public function down()
     {
         Schema::table($this->getUserTableName(), function (Blueprint $table) {
-            $table->dropColumn('verified');
-            $table->dropColumn('verification_token');
+            if (Schema::hasColumn($this->getUserTableName(), 'verified'))
+            {
+                Schema::table($this->getUserTableName(), function(Blueprint $table)
+                {
+                    $table->dropColumn('verified');
+                });
+            }
+            if (Schema::hasColumn($this->getUserTableName(), 'verification_token'))
+            {
+                Schema::table($this->getUserTableName(), function(Blueprint $table)
+                {
+                    $table->dropColumn('verification_token');
+                });
+            }
         });
     }
 }
